+++ superwall/src/main/java/com/superwall/sdk/config/ConfigManager.kt
@@ -1,15 +1,16 @@
 package com.superwall.sdk.config
 
-import LogLevel
-import LogScope
-import Logger
 import android.content.Context
+import android.webkit.WebView
 import com.superwall.sdk.config.models.ConfigState
 import com.superwall.sdk.config.models.getConfig
 import com.superwall.sdk.config.options.SuperwallOptions
 import com.superwall.sdk.dependencies.DeviceInfoFactory
 import com.superwall.sdk.dependencies.RequestFactory
 import com.superwall.sdk.dependencies.RuleAttributesFactory
+import com.superwall.sdk.logger.LogLevel
+import com.superwall.sdk.logger.LogScope
+import com.superwall.sdk.logger.Logger
 import com.superwall.sdk.misc.Result
 import com.superwall.sdk.misc.awaitFirstValidConfig
 import com.superwall.sdk.models.assignment.AssignmentPostback
@@ -249,40 +250,43 @@ open class ConfigManager(
 
     // Preloads paywalls referenced by triggers.
     private suspend fun preloadPaywalls(paywallIdentifiers: Set<String>) {
-        coroutineScope {
-            // List to hold all the Deferred objects
-            val tasks = mutableListOf<Deferred<Any>>()
+        val webviewExists = WebView.getCurrentWebViewPackage() != null
+        if (webviewExists) {
+            coroutineScope {
+                // List to hold all the Deferred objects
+                val tasks = mutableListOf<Deferred<Any>>()
 
-            for (identifier in paywallIdentifiers) {
-                val task = async {
-                    // Your asynchronous operation
-                    val request = factory.makePaywallRequest(
-                        eventData = null,
-                        responseIdentifiers = ResponseIdentifiers(
-                            paywallId = identifier,
-                            experiment = null
-                        ),
-                        overrides = null,
-                        isDebuggerLaunched = false,
-                        presentationSourceType = null,
-                        retryCount = 6
-                    )
-                    try {
-                        paywallManager.getPaywallViewController(
-                            request = request,
-                            isForPresentation = true,
-                            isPreloading = true,
-                            delegate = null
+                for (identifier in paywallIdentifiers) {
+                    val task = async {
+                        // Your asynchronous operation
+                        val request = factory.makePaywallRequest(
+                            eventData = null,
+                            responseIdentifiers = ResponseIdentifiers(
+                                paywallId = identifier,
+                                experiment = null
+                            ),
+                            overrides = null,
+                            isDebuggerLaunched = false,
+                            presentationSourceType = null,
+                            retryCount = 6
                         )
-                    } catch (e: Exception) {
-                        // Handle exception
+                        try {
+                            paywallManager.getPaywallViewController(
+                                request = request,
+                                isForPresentation = true,
+                                isPreloading = true,
+                                delegate = null
+                            )
+                        } catch (e: Exception) {
+                            // Handle exception
+                        }
                     }
+                    tasks.add(task)
                 }
-                tasks.add(task)
-            }
 
-            // Await all tasks
-            tasks.forEach { it.await() }
+                // Await all tasks
+                tasks.forEach { it.await() }
+            }
         }
     }