+++ superwall/src/main/java/com/superwall/sdk/identity/IdentityManager.kt
@@ -1,12 +1,12 @@
 package com.superwall.sdk.identity
 
-import LogLevel
-import LogScope
-import Logger
 import com.superwall.sdk.Superwall
 import com.superwall.sdk.analytics.internal.track
 import com.superwall.sdk.analytics.internal.trackable.InternalSuperwallEvent
 import com.superwall.sdk.config.ConfigManager
+import com.superwall.sdk.logger.LogLevel
+import com.superwall.sdk.logger.LogScope
+import com.superwall.sdk.logger.Logger
 import com.superwall.sdk.misc.awaitFirstValidConfig
 import com.superwall.sdk.misc.sha256MappedToRange
 import com.superwall.sdk.network.device.DeviceHelper