+++ superwall/src/main/java/com/superwall/sdk/logger/LogScope.kt
@@ -1,3 +1,5 @@
+package com.superwall.sdk.logger
+
 enum class LogScope {
     localizationManager,
     bounceButton,