+++ superwall/src/main/java/com/superwall/sdk/store/ExternalNativePurchaseController.kt
@@ -4,17 +4,26 @@ import android.app.Activity
 import android.content.Context
 import com.android.billingclient.api.*
 import com.superwall.sdk.Superwall
+import com.superwall.sdk.billing.RECONNECT_TIMER_MAX_TIME_MILLISECONDS
+import com.superwall.sdk.billing.RECONNECT_TIMER_START_MILLISECONDS
 import com.superwall.sdk.delegate.PurchaseResult
 import com.superwall.sdk.delegate.RestorationResult
 import com.superwall.sdk.delegate.SubscriptionStatus
 import com.superwall.sdk.delegate.subscription_controller.PurchaseController
+import com.superwall.sdk.logger.LogLevel
+import com.superwall.sdk.logger.LogScope
+import com.superwall.sdk.logger.Logger
+import com.superwall.sdk.store.abstractions.product.OfferType
+import com.superwall.sdk.store.abstractions.product.RawStoreProduct
 import com.superwall.sdk.store.abstractions.product.StoreProduct
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
+import kotlin.math.min
 
 class ExternalNativePurchaseController(var context: Context) : PurchaseController, PurchasesUpdatedListener {
     private var billingClient: BillingClient = BillingClient.newBuilder(context)
@@ -23,6 +32,8 @@ class ExternalNativePurchaseController(var context: Context) : PurchaseControlle
         .build()
     private val isConnected = MutableStateFlow(false)
     private val purchaseResults = MutableStateFlow<PurchaseResult?>(null)
+    // how long before the data source tries to reconnect to Google play
+    private var reconnectMilliseconds = RECONNECT_TIMER_START_MILLISECONDS
 
     //region Initialization
 
@@ -33,20 +44,43 @@ class ExternalNativePurchaseController(var context: Context) : PurchaseControlle
     }
 
     private fun startConnection() {
-        billingClient.startConnection(object : BillingClientStateListener {
-            override fun onBillingSetupFinished(billingResult: BillingResult) {
-                isConnected.value = billingResult.responseCode == BillingClient.BillingResponseCode.OK
-                syncSubscriptionStatus()
-            }
-
-            override fun onBillingServiceDisconnected() {
-                isConnected.value = false
-
-                CoroutineScope(Dispatchers.IO).launch {
-                    startConnection()
+        try {
+            billingClient.startConnection(object : BillingClientStateListener {
+                override fun onBillingSetupFinished(billingResult: BillingResult) {
+                    isConnected.value =
+                        billingResult.responseCode == BillingClient.BillingResponseCode.OK
+                    syncSubscriptionStatus()
                 }
-            }
-        })
+
+                override fun onBillingServiceDisconnected() {
+                    isConnected.value = false
+
+                    Logger.debug(
+                        LogLevel.error,
+                        LogScope.nativePurchaseController,
+                        "ExternalNativePurchaseController billing client disconnected, " +
+                                "retrying in $reconnectMilliseconds milliseconds",
+                    )
+
+                    CoroutineScope(Dispatchers.IO).launch {
+                        delay(reconnectMilliseconds)
+                        startConnection()
+                    }
+
+                    reconnectMilliseconds = min(
+                        reconnectMilliseconds * 2,
+                        RECONNECT_TIMER_MAX_TIME_MILLISECONDS,
+                    )
+                }
+            })
+        } catch (e: IllegalStateException) {
+            Logger.debug(
+                LogLevel.error,
+                LogScope.nativePurchaseController,
+                "IllegalStateException when connecting to billing client for " +
+                        "ExternalNativePurchaseController: ${e.message}",
+            )
+        }
     }
 
     //endregion
@@ -63,16 +97,37 @@ class ExternalNativePurchaseController(var context: Context) : PurchaseControlle
 
     //region PurchaseController
 
+    private fun buildFullId(
+        subscriptionId: String,
+        basePlanId: String?,
+        offerId: String?
+    ): String =
+        buildString {
+            append(subscriptionId)
+            basePlanId?.let { append(":$it") }
+            offerId?.let { append(":$it") }
+        }
+
     override suspend fun purchase(
         activity: Activity,
         productDetails: ProductDetails,
         basePlanId: String?,
         offerId: String?
     ): PurchaseResult {
-        val offerToken = productDetails.subscriptionOfferDetails
-            ?.firstOrNull { it.offerId == offerId }
-            ?.offerToken
-            ?: ""
+        val fullId = buildFullId(
+            subscriptionId = productDetails.productId,
+            basePlanId = basePlanId,
+            offerId = offerId
+        )
+
+        val rawStoreProduct = RawStoreProduct(
+            underlyingProductDetails = productDetails,
+            fullIdentifier = fullId,
+            basePlanId = basePlanId,
+            offerType = offerId?.let { OfferType.Offer(id = it) }
+        )
+
+        val offerToken = rawStoreProduct.selectedOffer?.offerToken ?: ""
 
         val productDetailsParams = BillingFlowParams.ProductDetailsParams.newBuilder()
             .setProductDetails(productDetails)
@@ -156,7 +211,7 @@ class ExternalNativePurchaseController(var context: Context) : PurchaseControlle
         val hasActivePurchaseOrSubscription = allPurchases.any { it.purchaseState == Purchase.PurchaseState.PURCHASED }
         val status: SubscriptionStatus = if (hasActivePurchaseOrSubscription) SubscriptionStatus.ACTIVE else SubscriptionStatus.INACTIVE
 
-        if (Superwall.initialized == false) {
+        if (!Superwall.initialized) {
             Logger.debug(
                 logLevel = LogLevel.error,
                 scope = LogScope.nativePurchaseController,